# services/soup-identity/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Soup Identity API
  version: 1.0.0
  description: |
    Thin Soup identity microservice (v1 frozen).
    Trust rules: Soup never calls Jellyfin; no QC mailbox; multi-server is always `servers[]`.
servers:
  - url: http://localhost:8787
    description: Local docker-compose / `npm run dev`
...
security:
  - SoupBearer: []
paths:
...
  /auth/device/link/{deviceCode}:
    get:
      tags: [auth]
      summary: Poll device-link until approved
      security: []
      parameters:
        - $ref: "#/components/parameters/deviceCode"
      responses:
        "200":
          description: Pending, expired, or completed with tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceLinkPoll"
        "404":
          description: Unknown device code
  /auth/device/approve:
    post:
      tags: [auth]
      summary: Approve a device-link user code after Google (or dev) identity
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceLinkApprove"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [approved]
        "400":
          description: Invalid or expired code
        "404":
          description: Unknown user code
  /auth/google/start:
    get:
      tags: [auth]
      summary: Begin Google OIDC (browser). Query user_code binds to device-link.
      security: []
      parameters:
        - name: user_code
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to Google (or local approve page when ALLOW_DEV_LOGIN)
        "503":
          description: Google client not configured and dev login disabled
  /auth/google/callback:
    get:
      tags: [auth]
      summary: Google OIDC callback
      description: >
        On success, auto-approves when OAuth state carries user_code; otherwise
        returns an HTML form to enter the TV code (signed link_token proof).
        Never returns Soup session tokens in the browser.
      security: []
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: error
          in: query
          schema:
            type: string
      responses:
        "200":
          description: HTML success or enter-TV-code form
        "400":
          description: OAuth error or expired TV code (may still show enter-code form)
  /auth/google/link-device:
    post:
      tags: [auth]
      summary: Finish Google device-link with the TV user code
      description: >
        Browser form post after Google SSO. Requires link_token (short-lived
        signed proof from the callback) and user_code from the TV.
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [link_token, user_code]
              properties:
                link_token:
                  type: string
                user_code:
                  type: string
      responses:
        "200":
          description: HTML success — TV can poll for tokens
        "400":
          description: Invalid/expired proof or TV code (re-shows form)
  /auth/dev/complete:
    post:
      tags: [auth]
      summary: Dev-only approve helper when Google client env is unset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_code]
              properties:
                user_code:
                  type: string
                email:
                  type: string
      responses:
        "200":
          description: Approved via DEV_GOOGLE_EMAIL default or body override
        "403":
          description: Dev login disabled
  /v1/sessions/refresh:
    post:
      tags: [sessions]
      summary: Exchange refresh token for new access + rotated refresh
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: New session tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionTokens"
        "401":
          description: Invalid or revoked refresh
  /v1/sessions/revoke:
    post:
      tags: [sessions]
      summary: Revoke current refresh session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        "204":
          description: Revoked (idempotent)
        "401":
          description: Unauthorized
  /v1/me:
    get:
      tags: [me]
      summary: Current Soup subject
      responses:
        "200":
          description: Subject profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
        "401":
          description: Unauthorized
  /v1/me/servers:
    get:
      tags: [me]
      summary: Entitled servers roster (always list-shaped)
      description: >
        Atomically claims any claimable transport grant per server and returns
        material once in transport_grant; later reads return null for that grant.
      responses:
        "200":
          description: Roster
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerRoster"
        "401":
          description: Unauthorized
  /v1/assertions:
    post:
      tags: [assertions]
      summary: Mint short-lived JWT assertion for a target server (Pattern A)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [server_id]
              properties:
                server_id:
                  type: string
      responses:
        "200":
          description: Assertion JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssertionResponse"
        "401":
          description: Unauthorized
        "403":
          description: Not entitled to server
        "404":
          description: Unknown server
...
  /v1/me/servers/{serverId}/transport-grant/claim:
    post:
      tags: [app]
      summary: Claim a single-use transport grant for one server
      description: >
        Atomic single-claim. Equivalent semantics to the claim performed by
        GET /v1/me/servers. Returns material once; subsequent claims are 410.
      security:
        - SoupBearer: []
      parameters:
        - $ref: "#/components/parameters/serverId"
      responses:
        "200":
          description: Claimed grant with material
          content:
            application/json:
              schema:
                type: object
                required: [transport_grant]
                properties:
                  transport_grant:
                    $ref: "#/components/schemas/TransportGrantRecord"
        "403":
          description: Not entitled
        "410":
          description: No claimable grant (expired, claimed, revoked, or missing)
components:
  securitySchemes:
    SoupBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived Soup access token from device-link or refresh
    PluginBasic:
      type: http
      scheme: basic
      description: plugin_id:plugin_secret issued at plugin setup / BOOTSTRAP_PLUGIN_*
  parameters:
    deviceCode:
      name: deviceCode
      in: path
      required: true
      schema:
        type: string
    serverId:
      name: serverId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    DeviceLinkPoll:
      oneOf:
        - type: object
          required: [status]
          properties:
            status:
              type: string
              enum: [pending, expired]
        - allOf:
            - type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [approved]
            - $ref: "#/components/schemas/SessionTokens"
    DeviceLinkApprove:
      type: object
      required: [user_code, email]
      properties:
        user_code:
          type: string
        email:
          type: string
          description: Normalized Google account email
        name:
          type: string
    SessionTokens:
      type: object
      required: [access_token, refresh_token, expires_in, token_type, email]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        token_type:
          type: string
          enum: [Bearer]
        email:
          type: string
          description: Normalized Google account email
    Subject:
      type: object
      required: [email]
      properties:
        email:
          type: string
          description: Normalized Google account email
        name:
          type: string
          nullable: true
        picture_url:
          type: string
          nullable: true
    ServerRoster:
      type: object
      required: [servers]
      properties:
        servers:
          type: array
          items:
            $ref: "#/components/schemas/RosterServer"
    RosterServer:
      type: object
      required: [id, name, audience]
      properties:
        id:
          type: string
        name:
          type: string
        base_url:
          type: string
          nullable: true
        magic_dns:
          type: string
          nullable: true
        audience:
          type: string
        entitlement:
          type: object
          additionalProperties: true
        transport_grant:
          description: >
            On first successful roster read, Soup atomically claims any claimable
            grant and returns material once. Subsequent reads return null.
          nullable: true
          allOf:
            - $ref: "#/components/schemas/TransportGrantRecord"
    AssertionResponse:
      type: object
      required: [assertion, expires_in, server_id, audience]
      properties:
        assertion:
          type: string
          description: Compact JWT (sub=Google email, aud=server audience, exp 2–5m)
        expires_in:
          type: integer
        server_id:
          type: string
        audience:
          type: string
...
    TransportGrantRecord:
      type: object
      required: [id, grant_type, expires_at]
      properties:
        id:
          type: string
          format: uuid
        grant_type:
          type: string
        expires_at:
          type: string
          format: date-time
        claimed_at:
          type: string
          format: date-time
          nullable: true
        material:
          type: string
          description: Only returned once on claim; omitted from plugin list
        tailscale_key_id:
          type: string
          nullable: true
        capabilities:
          type: object
          additionalProperties: true